remove some more duplication
